* master: (37 commits)
  #73594 tests only check the extra params if dns_get_record is successful
  Fixed bug #74852 (property_exists returns true on unknown DateInterval property)
  fix uninitialized var
  fix comparison warning
  comply with POSIX signature
  fix warning
  remove some casts
  cleanup casts
  remove useless cast
  eliminate casts
  sync vim mode lines in main
  [ci skip] update NEWS
  [ci skip] update NEWS
  [ci skip] update NEWS
  Fixed bug #74883 SQLite3::__construct() produces "out of memory" exception with invalid flags
  Silent compiler warning
  Fix test
  Deprecated the read_exif_data() alias
  Add myself as exif maintainer
  update libs versions
  ...

+++ main/php_open_temporary_file.c
@@ -114,7 +114,7 @@ static int php_do_open_temporary_file(const char *path, const char *pfx, zend_st
 	}
 
 #ifdef PHP_WIN32
-	if (!php_win32_check_trailing_space(pfx, (const int)strlen(pfx))) {
+	if (!php_win32_check_trailing_space(pfx, strlen(pfx))) {
 		SetLastError(ERROR_INVALID_NAME);
 		return -1;
 	}
